set long phno to Student model and dto

+++ src/main/java/com/springboot_crud/springboot_crud/Dtos/StudentDto.java
@@ -14,6 +14,6 @@ public class StudentDto {
     private String email;
     private String department;
     private String gpa;
-    private int phno;
+    private long phno;
     private String address;
 }
